requirements.txt
mmpy_bot>=2.0
openstacksdk
click
schema
prometheus-api-client
